Factor out color-util.ts.
